requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycoingecko==1.3.0
pytz==2020.1
requests==2.24.0
selenium==3.141.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
